+++ CHANGELOG.md
@@ -2,16 +2,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.x.x] - 2021-xx-xx
+## [0.x.x] - 2022-xx-xx
 * Updated `cosmic_gps` to support xarray Datasets
 * Added altitude binning profile support for all datasets with altitude
   as a variable.
 * Added `download` to general methods
+* Added support for `cosmic2_ivm`
 * Maintenance
   * Update links in documentation
   * Update testing compliance
   * Add windows testing
   * Add workflow for testing with pysat RC
+  * Updated instrument tests to use pysat 3.0.2 syntax
+  * Added clean warning routine to `methods.general`
 
 ## [0.0.2] - 2021-06-18
 * Update instrument style for pysat 3.0.0